# students/k3340/laboratory_works/Bogdanov_Maxim/Lr3/docs/swagger/swagger.yaml
basePath: /api/v1
definitions:
...
  internal_delivery_http_handler.CreateStudentRequest:
    properties:
      class_id:
        type: integer
      first_name:
        type: string
      gender_id:
        type: integer
      last_name:
        type: string
      middle_name:
        type: string
    type: object
...
  internal_delivery_http_handler.StudentResponse:
    properties:
      class_id:
        type: integer
      created_at:
        type: string
      first_name:
        type: string
      gender_id:
        type: integer
      id:
        type: integer
      last_name:
        type: string
      middle_name:
        type: string
      updated_at:
        type: string
    type: object
...
  internal_delivery_http_handler.UpdateStudentRequest:
    properties:
      class_id:
        type: integer
      first_name:
        type: string
      gender_id:
        type: integer
      last_name:
        type: string
      middle_name:
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@school-service.com
    name: API Support
  description: API для управления школьной системой
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: School Management Service API
  version: "1.0"
paths:
...
  /students:
    get:
      consumes:
      - application/json
      description: Возвращает список всех учеников
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/internal_delivery_http_handler.StudentResponse'
            type: array
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Список учеников
      tags:
      - students
    post:
      consumes:
      - application/json
      description: Создает нового ученика в системе
      parameters:
      - description: Данные ученика
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_delivery_http_handler.CreateStudentRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/internal_delivery_http_handler.StudentResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Создать ученика
      tags:
      - students
  /students/{id}:
    delete:
      consumes:
      - application/json
      description: Выполняет soft delete ученика
      parameters:
      - description: ID ученика
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Удалить ученика
      tags:
      - students
    get:
      consumes:
      - application/json
      description: Возвращает информацию об ученике по его ID
      parameters:
      - description: ID ученика
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_delivery_http_handler.StudentResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Получить ученика по ID
      tags:
      - students
    put:
      consumes:
      - application/json
      description: Обновляет информацию об ученике
      parameters:
      - description: ID ученика
        in: path
        name: id
        required: true
        type: integer
      - description: Данные ученика
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_delivery_http_handler.UpdateStudentRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_delivery_http_handler.StudentResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Обновить ученика
      tags:
      - students
  /students/class/{classId}:
    get:
      consumes:
      - application/json
      description: Возвращает список всех учеников указанного класса
      parameters:
      - description: ID класса
        in: path
        name: classId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/internal_delivery_http_handler.StudentResponse'
            type: array
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Список учеников класса
      tags:
      - students
...
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"